fix: npm install errors with node-sass
Fix npm install errors with node-sass and gyp.

Switch from node-sass to dart-sass aka just "sass" now.
Update some aspects of gulp and the gulp script.

For reference, here is my machine and node environment info.
(Node is probably most relevant, but for whatever reason node-sass was
failing to build even when I switched to  Node 4.2.1 or Node 4.5.0 so it
made me wonder if gyp and node-sass just hate MacOS Ventura?)
- Machine: Macbook Pro 2018 (Intel Chip) MacOS Ventura  13.2.1 (22D68)
- Node v19.7.0, npm 9.5.0

As of this commit, with the above environment on my machine,
`npm install` and `npm run serve` work correctly.